auth_service/utils/token.py
import jwt
import datetime


SECRET_KEY = "B4Nu_Jhi3ftX10psMdWDFbYSmESqKmZILmCAH0zpAIs"
ALGORITHM = "HS256"
ACCESS_TOKEN_EXPIRE_MINUTES = 30


def create_jwt_token(username, role):
    payload = {
        "sub": username,
        "role": role,
        "exp": datetime.datetime.utcnow() + datetime.timedelta(hours=2)
    }
    return jwt.encode(payload, SECRET_KEY, algorithm="HS256")
