requirements.txt
grpcio 
grpcio-tools
sqlalchemy 
psycopg2-binary
protobuf
pyjwt
passlib[bcrypt]
